requirements.txt
Django>=1.10,<1.11
django-environ==0.4.0
django-smartcc>=0.1,<0.2
django-markdownx==1.6.3
pytz==2016.6.1
python-akismet==0.2.3
